web/requirements.txt
Flask
Flask-SocketIO
python-socketio
eventlet
